requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.0
asyncpg==0.30.0
autopep8==2.3.1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
debugpy==1.8.7
dnspython==2.7.0
docker==7.1.0
dotenv-python==0.0.1
dump-env==1.5.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==30.8.2
fastapi==0.115.4
fastapi-users==14.0.0
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
loguru==0.7.2
makefun==1.15.6
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
polyfactory==2.17.0
psycopg2-binary==2.9.10
pwdlib==0.2.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-alembic==0.11.1
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
python-telegram-bot==21.7
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.41.2
testcontainers==4.8.2
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wheel==0.44.0
wrapt==1.16.0
